Add ExampleMarkerDirective
This directive marks the scope of example content in the original
documentation, and lets authors add a title and tags.

Currently a pass-through directive. It parses the content of the
directive and adds it back to the document

Examples are persisted in the build environment for later
post-processing (to build the example gallery). Examples are keyed by
their unique ID (a slugified version of the title). The dict items
contains metadata and the content of the example (to later generate
standalone example pages).

The directive also collects title and tags as metadata.