Set up testing with sphinx.testing.fixtures
sphinx.testing.fixtures let us build Sphinx sites from pytest and then
inspect the built site. There can be multiple test sites, each test site
is a directory in the sphinx_astropy/test/roots/ directory.

This is the same pattern that Sphinx uses for its test, so this is likely
the easiest way to test our Sphinx extensions.
http://www.sphinx-doc.org/en/master/devguide.html#unit-testing

Note I had to add the pytest_plugins line, to load the Sphinx pytest
plugin, from a new conftest.py file at the root of the project, not from
sphinx_astropy/tests/conftest.py The reason for this is outlined in
https://docs.pytest.org/en/latest/deprecations.html#pytest-plugins-in-non-top-level-conftest-files

The rest of the pytest configuration is done in
sphinx_astropy/tests/conftest.py, which is where you'd expect most
configuration to go. This configuration is largely based on Sphinx's:
https://github.com/sphinx-doc/sphinx/blob/master/tests/conftest.py