Test that example marker creates target nodes
This test generates a site in the XML format since then it's easy to
search for nodes and their attributes.

Unfortunately this test strategy doesn't work for Sphinx <1.7, because
the pytest fixtures aren't available. Thus I have pytest skip these
tests for Sphinx <1.7.  I think this is still the best way to test
sphinx extensions and will continue to be so in the future because this
is how Sphinx tests itself.