Enable debug-level logging in tests
If the tests are failing, it's useful to see the debug-level logging.

Note: the '2' verbosity enables DEBUG-level logging. I can't find a
cleaner alias to this.

Also, it would be nicer to make this the default while using the sphinx
pytest mark, but I can't find an easy way to do that.