Remove examples in env-purge-doc event
The purge_doc callback is required to remove examples from a cached
environment if a document is removed in a subsequent build. Otherwise
the examples in the cached environment from previous builds would
continue to exist in subsequent builds.

The tests simulate a env-purge-doc event and separately ensure that
purge_examples got registered as a env-purge-doc callback.