Distinguish example ID from ref ID
These should be separate things so that the "example ID" is the
slugified version of the title (which is unique by design). Then the ref
ID has the example-src prefix to be a unique reference ID to the
example's source location. Also adds a ref_id field to the example's
dict in the build environment.