Implement Jinja-based templates
Templates for the standalone example pages, landing pages, tag browsing
pages, and so on, can be Jinja-formatted templates. This implementation
is adapted from sphinx.ext.autosummary, which has similar needs. By
doing this, the user can customize the templates at the builder, theme,
or project level.

The render is integrated into the ExamplePage class
(ExamplePage.render), which automatically detects and uses a template
named 'astropy_example/examplepage.rst'. The extension ships with a
default implementation of the template.

The under tests demonstrate rendering a standalone example page. This
ins't hooked up to piping in the content of the example, yet, though.