Write stubs for example pages in builder-inited
This commit puts together the work on detecting examples from source
(detect_examples) and the work on rendering stubs for standalone example
pages (ExamplePage) and run a pipeline during the builder-inited event.
This event happens early in the Sphinx build processor so that we can
create example pages before Sphinx actually begins to read and process
these pages.

This also adds a new config variable, astropy_examples_h1, which
customizes the underline character used for making titles for "h1"
headings in reStructuredText.