Render a basic gallery landing page
The landing page is the index.rst for the example gallery. It provides a
toctree for all the examples. The LandingPage class is implemented
similarly to the ExamplePage class in that it takes page data and is
responsible for computing paths, docnames, and rendering for itself.

In the future the template for the landing page could be enhanced into a
tiled gallery view, for example.

The test verifies that the index.rst file's reStructuredText is rendered
correctly.

Since there's now a toctree, individual example pages don't need the
`:orphan:` field.