Add TagPage for indexing examples for given tag
The TagPage is like a specialized version of the LandingPage that
indexes exaples that have a given tag.

The TagPage.generate_tag_pages constructor simultaneously makes tag
pages for the set of tags given the population of examples, and also
provides references to those tag pages with each relevant example page.