Add is_directive_registered utility
This lets us check if a directive is available. It will become useful
for the doctest skipping, later. It's also useful as a check of the
extension setup (the function is tested by also testing that the
example and example-content directives got set up).

Note that this function is available in Sphinx 2.0 and later, so this
version can be retired once that becomes the minimum required version.