Adapt image nodes in examples
This is similar to how pending_xref nodes need to be adapted if they use
a relative targetref. In this case, we adapt the uri attribute of the
image node to be an absolute URI in case a relative image path is used.
Absolute URIs and external (http) URIs are ignored.

Both the relative/local and external cases are tested here.