Adapt reference nodes generated by matplotlib
Matplotlib's "plot" directive uses external links to reference the code
and image files. This is because the matplotlib extension is rendering
out-of-band with the main Sphinx build, so "ref", "doc", and "download"
roles don't work.

The trick, then, is to look for external links (a reference node) and
process refuri attributes that start with ".//", and convert those into
relative URLs.