Add astropy_examples_enabled configuration
The default for this configuration is **False**. This is done so that
the extension can be enabled in the default Sphinx configuration for
packages, and those packages can begin using the example directive. They
won't get an example gallery though, which they may not be ready for
(you need to link to the gallery index page, for instance).

This configuration then lets a package opt-in to generating the example
gallery itself.

I think this will provide a smoother adoption experience.