Skip additional tests with Sphinx 1.7
Named equation reference links do not seem to work with Sphinx 1.7.

As well, the format for download links from the Matplotlib plot
extension with Sphinx 1.7 are different compared to more recent Sphinx
versions, so its best to skip that test since the tests would need to be
customized for Sphinx 1.7.