Copy substitution defs in ExampleMarkerDirective
Since substitution_reference nodes are resolved just _after_ the
ExampleMarkerDirective is run, substitution_reference nodes could be part
of the example content that is republished on the standalone example
page. Thus we also need to copy the substitution_definition nodes to
include them in the standalone example page.

Based on experimentation, it seems that traversing the document's nodes
and the directive content's nodes together only gets substitution
definitions that are written above or within the example directive. This
is a caveat that will need to be added to the documentation.