Example marker wraps example in custom node
Now the ExampleMarkerDirective wraps the generated example in a custom
container node, ExampleMarkerNode.

This node has visitors that implement this node into a <div> in HTML
with a class of astropy-example-source. This div+class marks the content
of an example so that it can be copied by the HTML postprocessor into
the standalone example pages.

NOTE: is_node_registered() is backported here from Sphinx 1.8+

This small function from is needed for tests, and is no longer needed
here if Sphinx 1.8 becomes the minimum supported version of the
extension.