ExampleContentDirective inserts a marker node
This node, ExampleContentNode, just provides an empty <div> on the HTML
page.

The class is "astropy-example-content" and the ID is the ID of the
example (which will be used to look up the example content).

Because we're no longer adding example content into the standalone
example page as part of the regular Sphinx/docutils build, we no longer
need to access and process any of the docutils nodes related to the
original example content.

Many tests are skipped because the expected content isn't be rendered
given this change. These tests will be reactivated later when examples
are being published via an HTML postprocessor.