Examples no longer stored in BuildEnvironment
Content nodes and metadata are no longer stored in the
`sphinx_astropy_examples` attribute of the BuildEnvironment. Now that
exmaples are copied and rendered as an HTML post-processing step, we
don't need to keep this metadata in the Sphinx build. Consequently, some
additonal hooks can also go:

- merge_examples in env-merge-info
- purge_examples in env-purge-doc
- reorder_example_page_reading in env-before-read-docs

This also means that the extension can be marked as safe for reading in
parallel.

Some tests no longer apply because the
BuildEnvironment.sphinx_astropy_examples attribute is no longer set. In
many cases, I've marked these to be skipped; eventually we'll want to
reactivate these tests in the future given the new scheme.