Cache example metadata after preprocessing
Metadata is cached to the sphinx_astropy_examples attribute of the
BuildEnvironment as part of the example gallery preprocessing step
during build-inited (recall that before this caching got added as part
of the example directive.

This is necessary so that the postprocessing step know where to find the
source for each example and the associated standalone example page.

This is now also the best way to check if there are duplicate examples
since all examples are combined at this point.

Notes on tests:

1. It seems this SphinxError is build raised in the equivalent of the
test set up with the pytest.mark.sphinx pytest extension. This means
that a full build is the best way to simulate this.

2. Because docstrings are not scanned by the preprocessor, examples
embedded in docstrings are no longer part of the example gallery. Thus
I've dropped the associated test to look for example from the
example_func docstring.