Create standalone example pages in HTML
This implements a new approach for populating standalone example pages.
Rather than republish the examples to standalone example pages as part
of the Sphinx build, this approach operates on the HTML that Sphinx has
built.

The postprocess_examples function operates as a hook for the
build-finished Sphinx event. For every example in the cached
sphinx_astropy_examples attribute of the BuildEnvironment, this method
extracts the example from the source page, adapts any relative links and
image sources, and inserts the example's <div> into the standalone
example page.

If config.astropy_examples_enabled is False, then do not run post-processing.