Overhaul testing for html postprocessing scheme
With the large change in sphinx_astropy.ext.example to populate examples
by post processing HTML rather than copying docutils nodes, we had to
temporarily disable many of the unit tests. This commit re-works the
tests to work with the new processing strategy.

The pytest.mark.sphinx mark doesn't work with the build-finished event
(not sure why), but this means we can't use that approach to test the
HTML output. Instead, we use the CLI-based build of the example
projects. Since a lot of test functions consume those builds, I've
turned them into session-scoped fixtures.

Now that BeautifulSoup4 is a dependency of sphinx_astropy, the tests use
BeautifulSoup to check the build output. This approach is a lot cleaner
than directly using the built-in html.parser module.

Finally, I've also organized the tests around different types:

1. Unit tests that don't depend on a Sphinx build.
2. Unit tests that use the ``sphinx`` pytest mark. These tests operate on
   the Sphinx application instance after a build and test environment
   persistence.
3. Tests that run a Sphinx build through its command-line interface and
   analyze the resulting HTML product.