Purge examples directory on each Sphinx build
We want to support incremental rebuilds so that only those example on
pages that were changed are re-scanned and rebuilt. However, that
involves finer-grained cache invalidation. To make the example extension
work with incremental rebuilds we're starting with the *simplest* thing,
which is to start the example gallery fresh on each build. The
sphinx_astropy_examples attribute on the BuildEnvironment was already
being reset on each build; now the example directory in the source tree
itself is deleted and re-created on each Sphinx build.